+++ config/config.exs
@@ -123,6 +123,8 @@ end
 # Configure Mnesia directory (used by pow persistent sessions)
 config :mnesia, :dir, to_charlist(System.get_env("MNESIA_DIR", ".mnesia"))
 
+config :appsignal, :config, revision: System.get_env("SHA", default_sha)
+
 # Import environment specific config. This must remain at the bottom
 # of this file so it overrides the configuration defined above.
 import_config "#{Mix.env()}.exs"