+++ lib/oli/rendering/context.ex
@@ -7,7 +7,6 @@ defmodule Oli.Rendering.Context do
   defstruct user: nil,
             activity_map: %{},
             render_opts: %{
-              log_errors: true,
               render_errors: true
             },
             preview: false,